Added Simulated annealing support
The program is self contained in an header file, as we would heavily use
templating. The bitstring.hpp contains all the details of the custom
Hamiltonian and Simulated annealing class. Detailed documentation can
also be generated via DOxygen, and the make file now supports making the
documentation as well.

1. Added class for storing state in bits.
2. Added class to store Ising Hamiltonian and internally covert it to
QUBO form.
3. Added class that could perform one step of simulated annealing.