Removed templates
Since the program must run from the command line dynamically allocating
values for parallel tempering, templates could not serve properly for
this program. Templating is useful when the compiler knows the
parameters during compilation. In this case, we changed templates to
dynamic parameters.

This allows us to separate the header and code file.