Added save functionality
The states and energies computed by parallel tempering at every step can
now be stored. By default, the energies are always stored in
memory. Since storing states is memory intensive, they are only saved if
a special flag is set to true before annealing. The state file is
written compressed to disk with zlib in order to save space.