Compute parallel tempering on two instances
In order to find out when the parallel tempering has reached the Gibb's
distribution, we need to run two equivalent instances and check for the
overlap of these instances.

We have now added another set of bit strings in parallel tempering which
anneal along side the original instances.

We want to have an interface added to this object so that after each
annealing step, we can compute any arbitrary quantity along the anneal
path. To make everything simpler, we have moved out the save-to-file
procedures, which will now be implemented by a different class.