Added helper objects for parallel tempering.
In pthelper.hpp, we define an interface for computing objects that can
be pushed to parallel tempering process. This allows us to hook an
arbitrary computation to parallel tempering quantities which are
constructed from energies and state vector of the two parallel tempering
copies.

As an example, we implemented a helper object that writes compressed
data to memory keeping track of all energies and state vectors, and
then writes the compressed data to disk.