Added spin overlap calculation and plots
Spin overlap, or the Edwards-Anderson SG parameter helps us in deciding
when the simulation thermalises. Once the spin overlap reaches a
constant value, we can sample from that simulation and get an estimate
for the Gibbs distribution.

In order to make plots, we have also added CERN ROOT as a dependency,
though it would be nicer to use a light-weight alternative later.