Added helper class to determine thermalisation.
The simulation thermalises when the spin overlap of the two classes does
not changes any more. We can test this by calculating the derivative of
the mean spin overlap of two copies. A helper class can now keep track
of when this happens.

Also optimises data saving by adding a compression buffer.