Added autocorrelation routine and bug fixes
Once the system has themalised, we sample from the system. In order to
ensure that the sampling is fair, we compute the autocorrelation of the
energies and compute the autocorrelation length. For sampling form then
on, we skip instances so that the sampled instances are separated by
proper autocorrelation lengths.

Bug fixes:
1. Changed Makefile in order to handle different systems.
2. Helper objects can now be removed from algorithm as needed.
3. Replaced Boost circular buffer by a custom implemented buffer class
for cleaner dependencies.
4. Bit-strings are now randomised bit by bit to avoid buffer overflow of
boost bitset with number of qubits smaller than block size.
5. Thermalisation algorithm now checks that each beta has separately
thermalised.