Added routines to track spin overlap & visit freq.
In order to estimate that the simulation has truly thermalised, we need
to track the spin overlap to ensure that the histogram of observed spin
overlap values is symmetric. Also, we want to ensure that each instance
freely propagates through the temperature space. In order to confirm the
free propagation, we also tag the states and keep track of their
ordering after every swap.

Also fixed a bug in calculation of spin overlap, and added a function
which gives various information about the parallel tempering simulation
in order to asses whether the simulation thermalised or not.

The graphing library was changes to gnuplot.